Examined and visualize some categorical data
visualised the Year and Status attributes  and made some changes to the
Staus data values by changing PT to Part Time and FT to Full Time.
Also changed the year column from datetime to object type to be able
to visualise it.

I removed kde = False from the histogram plottings.